+++ packages/scaffold-config/src/dependencies.ts
@@ -31,7 +31,8 @@ export const WIZARD_DEPENDENCY_REACT = {
   package: 'react',
   installer: 'react',
   description: 'A JavaScript library for building user interfaces',
-  minVersion: '^16.0.0 || ^17.0.0 || ^18.0.0',
+  // FIXME: update to generic react once released
+  minVersion: '^16.0.0 || ^17.0.0 || ^18.0.0 || 19.0.0-rc-81c5ff2e04-20240521',
 } as const
 
 export const WIZARD_DEPENDENCY_REACT_DOM = {
@@ -40,7 +41,8 @@ export const WIZARD_DEPENDENCY_REACT_DOM = {
   package: 'react-dom',
   installer: 'react-dom',
   description: 'This package serves as the entry point to the DOM and server renderers for React',
-  minVersion: '^16.0.0 || ^17.0.0 || ^18.0.0',
+  // FIXME: update to generic react once released
+  minVersion: '^16.0.0 || ^17.0.0 || ^18.0.0 || 19.0.0-rc-81c5ff2e04-20240521',
 } as const
 
 export const WIZARD_DEPENDENCY_TYPESCRIPT = {