+++ system-tests/projects/react19/cypress/support/component.js
@@ -1,0 +1,5 @@
+import { mount } from 'cypress/react19'
+
+import './backgroundColor.css'
+
+Cypress.Commands.add('mount', mount)